Rename file back to kebab-case
