+++ apps/app/src/client/components/Admin/MarkdownSetting/ContentDispositionSettings.tsx
@@ -3,7 +3,7 @@ import React, { useState, useCallback, useEffect } from 'react';
 import { useTranslation } from 'next-i18next';
 import { useForm } from 'react-hook-form';
 
-import { useContentDisposition, type ContentDispositionSettings } from '../../../services/AdminContentDispositionSettings';
+import { useContentDisposition, type ContentDispositionSettings } from '../../../services/admin-content-disposition';
 import AdminUpdateButtonRow from '../Common/AdminUpdateButtonRow';
 
 interface MimeTypeListProps {
